chore(core): re exports modules and update imports
