+++ src/db.rs
@@ -2,7 +2,7 @@
 
 use crate::file_manager::FileManager;
 use crate::serialization::{Deserializer, serialize};
-use crate::trie::{error::TrieError, node::Node, node_hash::NodeHash};
+use crate::trie::{TrieError, Node, NodeHash};
 use std::path::PathBuf;
 
 /// Ethrex DB struct
@@ -158,7 +158,7 @@ impl EthrexDB {
 
 #[cfg(test)]
 mod tests {
-    use crate::trie::trie::{InMemoryTrieDB, Trie};
+    use crate::trie::{InMemoryTrieDB, Trie};
 
     use super::*;
     use tempdir::TempDir;
