+++ src/file_manager.rs
@@ -1,4 +1,4 @@
-use crate::trie::error::TrieError;
+use crate::trie::TrieError;
 use memmap2::{Mmap, MmapOptions};
 use std::fs::{File, OpenOptions};
 use std::io::{Seek, SeekFrom, Write};
