+++ src/trie/nibbles.rs
@@ -1,11 +1,6 @@
 use std::cmp;
 
-use crate::rlp::{
-    decode::RLPDecode,
-    encode::RLPEncode,
-    error::RLPDecodeError,
-    structs::{Decoder, Encoder},
-};
+use crate::rlp::{RLPDecode, RLPEncode, RLPDecodeError, Decoder, Encoder};
 
 /// Struct representing a list of nibbles (half-bytes)
 #[derive(Debug, Clone, Default, PartialEq, Eq, PartialOrd, Ord)]
